Refactor control flow related to filtering
Operator::apply allows us to abstract operator checks, so FilterCriteria
should use it. Filtering a specific column should occur inside
FilterCriteria, rather than at ContactPredicate.

TodoPredicate is abstracted out of FilterTodoCommand, with plans for
further abstactions of commonalities.

In summary, there are 3 levels of filtering:
- Operator: which handles and/or/nand/nor
- Column: which handles such filtering for a speciifc column of item
- Item: which ensures that the item is true only if all columns are true
