+++ src/main/java/seedu/address/logic/commands/read/FilterContactCommand.java
@@ -22,28 +22,28 @@ import seedu.address.model.contact.ContactPredicate;
 public class FilterContactCommand extends FilterCommand<Contact> {
 
     public static final String MESSAGE_USAGE = CONTACT_COMMAND_WORD + " " + COMMAND_WORD + ": Filters contacts based on"
-        + "specified criteria and displays them as a list with index numbers.\n"
-        + "Parameters: <col>/ [<op>:] <value(s)> [...]\n"
-        + "- <col>/ : Column to filter on ("
-        + PREFIX_CONTACT_NAME_LONG + ", "
-        + PREFIX_CONTACT_EMAIL_LONG + ", "
-        + PREFIX_CONTACT_ID_LONG + ", "
-        + PREFIX_CONTACT_COURSE_LONG + ", "
-        + PREFIX_CONTACT_GROUP_LONG + ", "
-        + PREFIX_CONTACT_TAG_LONG + ")\n"
-        + "- <op>: : Operator (and, or, nand, nor). Defaults to 'and' if not specified\n"
-        + "- <value(s)>: One or more values to filter by. Use quotes for values with spaces.\n"
-        + "Examples:\n"
-        + "1. " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " " + PREFIX_CONTACT_ID_LONG + " or: 12 13\n"
-        + "   Find students with ID 12 or 13.\n"
-        + "2. " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " " + PREFIX_CONTACT_NAME_LONG + "\"Darren Tan\" "
-        + PREFIX_CONTACT_COURSE_LONG + " CS1010S " + PREFIX_CONTACT_GROUP_LONG + "or: T01 T02 T03\n"
-        + "   Find contacts with \"Darren Tan\" in their name who"
-        + " enroll in course CS1010S and class T01, T02, or T03.\n"
-        + "3. " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " "
-        + PREFIX_CONTACT_NAME_LONG + "nand: \"My enemy\" Hater " + PREFIX_CONTACT_TAG_LONG + "and: handsome smart\n"
-        + "   Find contacts whose names do not contain \"My enemy\" and"
-        + " \"Hater\" and are tagged with both \"handsome\" and \"smart\".";
+            + "specified criteria and displays them as a list with index numbers.\n"
+            + "Parameters: <col>/ [<op>:] <value(s)> [...]\n"
+            + "- <col>/ : Column to filter on ("
+            + PREFIX_CONTACT_NAME_LONG + ", "
+            + PREFIX_CONTACT_EMAIL_LONG + ", "
+            + PREFIX_CONTACT_ID_LONG + ", "
+            + PREFIX_CONTACT_COURSE_LONG + ", "
+            + PREFIX_CONTACT_GROUP_LONG + ", "
+            + PREFIX_CONTACT_TAG_LONG + ")\n"
+            + "- <op>: : Operator (and, or, nand, nor). Defaults to 'and' if not specified\n"
+            + "- <value(s)>: One or more values to filter by. Use quotes for values with spaces.\n"
+            + "Examples:\n"
+            + "1. " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " " + PREFIX_CONTACT_ID_LONG + " or: 12 13\n"
+            + "   Find students with ID 12 or 13.\n"
+            + "2. " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " " + PREFIX_CONTACT_NAME_LONG + "\"Darren Tan\" "
+            + PREFIX_CONTACT_COURSE_LONG + " CS1010S " + PREFIX_CONTACT_GROUP_LONG + "or: T01 T02 T03\n"
+            + "   Find contacts with \"Darren Tan\" in their name who"
+            + " enroll in course CS1010S and class T01, T02, or T03.\n"
+            + "3. " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " "
+            + PREFIX_CONTACT_NAME_LONG + "nand: \"My enemy\" Hater " + PREFIX_CONTACT_TAG_LONG + "and: handsome smart\n"
+            + "   Find contacts whose names do not contain \"My enemy\" and"
+            + " \"Hater\" and are tagged with both \"handsome\" and \"smart\".";
 
     private final Predicate<Contact> predicate;
