+++ src/main/java/seedu/address/model/contact/FilterCriteria.java
@@ -1,45 +1,45 @@
 package seedu.address.model.contact;
 
 import java.util.List;
+import java.util.function.Predicate;
 
 import seedu.address.commons.core.Operator;
 
 /**
  * Represents the filter criteria for a single column.
- * Each criteria consists of a logical operator and a list of values to match against.
+ * Each criteria consist of a logical operator and a list of values to match against.
  */
 public class FilterCriteria {
     private final Operator operator;
-
-    private final List<String> values;
+    private final List<String> valuesToSearch;
 
     /**
      * Constructs a FilterCriteria with the given operator and values.
      *
-     * @param operator the logical operator to apply
-     * @param values the list of values to match against
+     * @param operator       the logical operator to apply
+     * @param valuesToSearch the list of values to match against
      */
-    public FilterCriteria(Operator operator, List<String> values) {
+    public FilterCriteria(Operator operator, List<String> valuesToSearch) {
         this.operator = operator;
-        this.values = values;
+        this.valuesToSearch = valuesToSearch;
     }
 
     /**
-     * Gets the logical operator for this criteria.
+     * Tests if the contact's values match the filter criteria.
+     * <p>
+     * Concrete example: if the criteria is for name with AND operator and values = ["A", "B"],
+     * and the contact's name is "Alice Bob", then the method will return true since both
+     * "A" and "B" are in the contact's name.
      *
-     * @return the operator
+     * @param itemColumnValues the values from the contact
+     * @return true if the values match the criteria, false otherwise
      */
-    public Operator getOperator() {
-        return operator;
-    }
-
-    /**
-     * Gets the list of values for this criteria.
-     *
-     * @return the list of values
-     */
-    public List<String> getValues() {
-        return values;
+    boolean testCriteria(List<String> itemColumnValues) {
+        // Checks if the value is contained within itemColumnValues in a case-insensitive manner
+        Predicate<String> isValueContained = searchValue -> itemColumnValues.stream().anyMatch(
+                itemValue -> itemValue.toLowerCase().contains(searchValue.toLowerCase())
+        );
+        return operator.apply(valuesToSearch.stream(), isValueContained);
     }
 
     @Override
@@ -53,14 +53,14 @@ public class FilterCriteria {
         }
 
         return operator == otherCriteria.operator
-                && values.equals(otherCriteria.values);
+                && valuesToSearch.equals(otherCriteria.valuesToSearch);
     }
 
     @Override
     public int hashCode() {
         int result = 17;
         result = 31 * result + operator.hashCode();
-        result = 31 * result + values.hashCode();
+        result = 31 * result + valuesToSearch.hashCode();
         return result;
     }
 }
